pyproject.toml
[project]
name = "langchain-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-identity>=1.25.1",
    "beautifulsoup4>=4.14.2",
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.1",
    "langchainhub>=0.1.21",
    "lxml>=6.0.2",
    "playwright>=1.55.0",
    "python-dotenv>=1.2.1",
    "reportlab>=4.4.4",
]
